ls: fix build failure on SunOS-5.10 (with libfts)

+++ ls/ls.c
@@ -38,6 +38,8 @@
 #include "mkc_getbsize.h"
 #include "mkc_bsd_getopt.h"
 #include "mkc_types.h"
+#include "mkc_err.h"
+#include "mkc_warn.h"
 
 #include <stdint.h>
 
@@ -61,7 +63,6 @@ __RCSID("$NetBSD: ls.c,v 1.74 2014/04/02 10:55:47 wiz Exp $");
 #include <sys/ioctl.h>
 
 #include <dirent.h>
-#include <err.h>
 #include <errno.h>
 #include <fts.h>
 #include <locale.h>
@@ -506,7 +507,7 @@ display(FTSENT *p, FTSENT *list)
 	DISPLAY d;
 	FTSENT *cur;
 	NAMES *np;
-	u_int64_t btotal, stotal;
+	uint64_t btotal, stotal;
 	off_t maxsize;
 	blkcnt_t maxblock;
 	ino_t maxinode;
@@ -590,14 +591,14 @@ display(FTSENT *p, FTSENT *list)
 				    (user = user_from_uid(sp->st_uid, 0)) ==
 				    NULL) {
 					(void)snprintf(nuser, sizeof(nuser),
-					    "%u", sp->st_uid);
+					    "%lu", (long unsigned)sp->st_uid);
 					user = nuser;
 				}
 				if (f_numericonly ||
 				    (group = group_from_gid(sp->st_gid, 0)) ==
 				    NULL) {
 					(void)snprintf(ngroup, sizeof(ngroup),
-					    "%u", sp->st_gid);
+					    "%lu", (long unsigned)sp->st_gid);
 					group = ngroup;
 				}
 				if ((ulen = strlen(user)) > maxuser)
